feat(cli): add storage implementation
