+++ cli/Cargo.toml
@@ -9,10 +9,16 @@ name = "ceres"
 path = "bin/ceres.rs"
 
 [dependencies]
+etc = "0.1.16"
 hex = "0.4.3"
+dirs = "3.0.1"
+sled = "0.34.6"
+serde = "1.0.125"
+bincode = "1.3.3"
 thiserror = "1.0.24"
 structopt = "0.3.21"
 
 # ceres deps
+ceres-std = { path = "../std", features = [ "std" ] }
 ceres-runtime = { path = "../runtime", features = [ "std" ] }
 ceres-sandbox = { path = "../sandbox", features = [ "std" ] }
