+++ cli/src/cmd/deploy.rs
