+++ cli/src/cmd/info.rs
